refactor: move maps query, sorting + nomap view into MapsList
