+++ src/components/MapsList/styles.ts
@@ -1,0 +1,6 @@
+import { styled } from '@mui/system'
+
+export const Img = styled('img')`
+  width: 350px;
+  height: 350px;
+`
