+++ src/views/Home/styles.ts
@@ -1,10 +1,5 @@
 import { styled } from '@mui/system'
 
-export const Img = styled('img')`
-  width: 350px;
-  height: 350px;
-`
-
 export const Overlay = styled('div')`
   width: 100vw;
   height: 100vh;
